+++ src/lib/HostProxy.h
@@ -29,9 +29,9 @@ public:
 	//return 0 in usb_ctrlrequest->brequest if there is no request
 	virtual int control_request(usb_ctrlrequest *setup_packet, int *nbytes, __u8** dataptr,int timeout=500)=0;
 
-	virtual void send_data(__u8 endpoint,__u8 attributes,__u16 maxPacketSize,__u8* dataptr,int length)=0;
+	virtual void send_data(__u8 endpoint,__u8 attributes,__u16 maxPacketSize,__u8* dataptr,__u64 length)=0;
 	virtual bool send_wait_complete(__u8 endpoint,int timeout=500) {return true;}
-	virtual void receive_data(__u8 endpoint,__u8 attributes,__u16 maxPacketSize,__u8** dataptr, int* length,int timeout=500)=0;
+	virtual void receive_data(__u8 endpoint,__u8 attributes,__u16 maxPacketSize,__u8** dataptr, __u64* length,int timeout=500)=0;
 	virtual void setConfig(Configuration* fs_cfg,Configuration* hs_cfg,bool hs)=0;
 	virtual char* toString() {return NULL;}
