+++ src/lib/Proxy.h
@@ -25,9 +25,9 @@ public:
 
 	virtual ~Proxy() {}
 
-	virtual void send_data(__u8 endpoint,__u8 attributes,__u16 maxPacketSize,__u8* dataptr,int length)=0;
+	virtual void send_data(__u8 endpoint,__u8 attributes,__u16 maxPacketSize,__u8* dataptr,__u64 length)=0;
 	virtual bool send_wait_complete(__u8 endpoint,int timeout=500) {return true;}
-	virtual void receive_data(__u8 endpoint,__u8 attributes,__u16 maxPacketSize,__u8** dataptr, int* length,int timeout=500)=0;
+	virtual void receive_data(__u8 endpoint,__u8 attributes,__u16 maxPacketSize,__u8** dataptr, __u64* length,int timeout=500)=0;
 	virtual void setConfig(Configuration* fs_cfg,Configuration* hs_cfg,bool hs)=0;
 	virtual char* toString() {return NULL;}
 	unsigned debugLevel;
